progress

+++ server/src/controller/manager-controller.ts
@@ -1,6 +1,8 @@
 import { Request, Response, NextFunction } from "express";
 import { PrismaClient } from "@prisma/client";
 import { RequestWithUserId } from "../types";
+import { createAdminManagerSchema } from "../schemas/admin-manager-schema";
+import { ZodError } from "zod";
 
 const prisma = new PrismaClient();
 
@@ -74,7 +76,8 @@ export async function createManager(
   next: NextFunction
 ) {
   try {
-    const { id } = req.body;
+    const parsedData = createAdminManagerSchema.parse(req.body);
+    const { id } = parsedData;
 
     const existingUser = await prisma.user.findUnique({
       where: {
@@ -127,6 +130,9 @@ export async function deleteManager(
     if (!existingManager)
       return res.status(404).json({ message: "Manager not found" });
 
+    if (id === process.env.SUPER_ADMIN_ID)
+      return res.status(403).json({ message: "Cannot delete Super Admin" });
+
     await prisma.manager.delete({
       where: {
         id: id,
@@ -134,6 +140,10 @@ export async function deleteManager(
     });
     return res.status(200).json({ message: "Manager deleted" });
   } catch (error) {
-    next(error);
+    if (error instanceof ZodError) {
+      return res.status(400).json({ errors: error.errors });
+    } else {
+      next(error);
+    }
   }
 }
